move WCStoreV1ProductsCollectionProps to hooks def
