+++ plugins/woocommerce-blocks/assets/js/blocks/product-collection/types.ts
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import type { BlockEditProps } from '@wordpress/blocks';
-import type { CurrencyCode, AttributeMetadata } from '@woocommerce/types';
+import type { AttributeMetadata } from '@woocommerce/types';
 
 /**
  * Internal dependencies
@@ -218,45 +218,3 @@ export type SetPreviewState = ( args: {
 	location: WooCommerceBlockLocation;
 	attributes: ProductCollectionAttributes;
 } ) => void | ( () => void );
-
-type AttributeCount = {
-	term: number;
-	count: number;
-};
-
-export type RatingValues = 0 | 1 | 2 | 3 | 4 | 5;
-
-type RatingCount = {
-	rating: RatingValues;
-	count: number;
-};
-
-type StockStatusCount = {
-	status: 'instock' | 'outofstock' | 'onbackorder';
-	count: number;
-};
-
-type PriceRangeProps = {
-	min_price: string;
-	max_price: string;
-	currency_code: CurrencyCode;
-	currency_decimal_separator: '.' | string;
-	currency_minor_unit: number;
-	currency_prefix: '$' | string;
-	currency_suffix: '' | string;
-	currency_symbol: '$' | string;
-	currency_thousand_separator: ',' | string;
-};
-
-/*
- * Prop types for the `wc/store/v1/products/collection-data` endpoint
- */
-export type WCStoreV1ProductsCollectionProps = {
-	price_range: PriceRangeProps | null;
-
-	attribute_counts: AttributeCount[] | null;
-
-	rating_counts: RatingCount[] | null;
-
-	stock_status_counts: StockStatusCount[] | null;
-};
